titan-poss-poss-web-f78a2c9ef984/titan-poss-poss-web-f78a2c9ef984/libs/shared/product-category-mapping/data-access-pc-mapping/src/lib/+state/pc-mapping.selector.ts
import { createSelector } from '@ngrx/store';
import { selectProductCategoryState } from './pc-mapping.reducer';

const selectProductCategory = createSelector(
  selectProductCategoryState,
  state => state.productCategory
);

const selectError = createSelector(
  selectProductCategoryState,
  state => state.error
);

const selectIsloading = createSelector(
  selectProductCategoryState,
  state => state.isLoading
);

export const ProductCategoryMappingSelectors = {
  selectProductCategory,
  selectError,
  selectIsloading
};
